Fix a bug where nbmpi breaks with >8 MPI tasks
The problem was that when splitting the data more than 8 ways, it
worked out so that one process did not have all of the classes.
Fixed by finding the possible classes beforehand from the entire
dataset and then passing those on to fit, rather than having fit
try to infer the classes.