Create a proof-of-concept for passing decision trees over MPI.
The program treempi trains a tree in one process and sends it to
another, which then applies it. We confirmed that we get the same
results through that method as by training and applying the tree
in the same process.

We currently use pickle to send the tree. Depending on performance,
we may want to investigate using MPI derived data types instead.