Fix bug with generator reuse in LazyDataSet
Some functions (e.g. split) attempt to reuse the generator of a
LazyDataSet multiple times (e.g. once for each split). This is
problematic, because once a generator is advanced it cannot be rewound,
so the second split doesn't get any cycles out of the generator.

Fixed by making the generator a function of no arguments that constructs
a new generator located at the beginning of the dataset.